requirements.txt
flask==3.1.0
flask-cors==5.0.0
langchain-huggingface==1.2.0
langchain-community==1.1.0
langchain-core==1.2.12
langchain-text-splitters==1.1.0
youtube-transcript-api==1.2.4
python-dotenv==1.2.1
faiss-cpu==1.9.0.post1
